+++ jetpack_vendor/automattic/jetpack-constants/CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.0.3-alpha] - unreleased
+
+This is an alpha version! The changes listed here are not final.
+
 ## [2.0.2] - 2024-04-30
 ### Changed
 - Internal updates.
@@ -166,6 +170,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Packages: Finish the constants package
 
+[2.0.3-alpha]: https://github.com/Automattic/jetpack-constants/compare/v2.0.2...v2.0.3-alpha
 [2.0.2]: https://github.com/Automattic/jetpack-constants/compare/v2.0.1...v2.0.2
 [2.0.1]: https://github.com/Automattic/jetpack-constants/compare/v2.0.0...v2.0.1
 [2.0.0]: https://github.com/Automattic/jetpack-constants/compare/v1.6.23...v2.0.0