+++ jetpack_vendor/automattic/jetpack-status/CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.2.2-alpha] - unreleased
+
+This is an alpha version! The changes listed here are not final.
+
+### Changed
+- Phab baseline file update.
+
 ## [3.2.1] - 2024-05-28
 ### Changed
 - Internal updates.
@@ -361,6 +368,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Packages: Introduce a status package
 
+[3.2.2-alpha]: https://github.com/Automattic/jetpack-status/compare/v3.2.1...v3.2.2-alpha
 [3.2.1]: https://github.com/Automattic/jetpack-status/compare/v3.2.0...v3.2.1
 [3.2.0]: https://github.com/Automattic/jetpack-status/compare/v3.1.0...v3.2.0
 [3.1.0]: https://github.com/Automattic/jetpack-status/compare/v3.0.3...v3.1.0