+++ jetpack_vendor/automattic/jetpack-waf/CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 This is an alpha version! The changes listed here are not final.
 
+### Changed
+- Phab baseline file update.
+
 ## [0.16.8] - 2024-05-20
 ### Changed
 - Internal updates.